// scent-quest-ui/src/app/services/models/page-response-review-response-picture.ts
/* tslint:disable */
/* eslint-disable */
import { ReviewResponsePicture } from '../models/review-response-picture';
export interface PageResponseReviewResponsePicture {
  content?: Array<ReviewResponsePicture>;
  first?: boolean;
  last?: boolean;
  number?: number;
  size?: number;
  totalElements?: number;
  totalPages?: number;
}